feature: parsing map and geting data of the .cub file

+++ lib/libft_plus/libft/ft_free_int_matrix.c
@@ -6,16 +6,16 @@
 /*   By: migarrid <migarrid@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/24 00:22:37 by migarrid          #+#    #+#             */
-/*   Updated: 2025/11/17 16:39:09 by migarrid         ###   ########.fr       */
+/*   Updated: 2026/01/02 04:07:45 by migarrid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft_plus.h"
 
-void	ft_free_int_matrix(int ***matrix_ptr, int rows)
+void	ft_free_int_matrix(int ***matrix_ptr, size_t rows)
 {
-	int	i;
-	int	**matrix;
+	size_t	i;
+	int		**matrix;
 
 	if (!matrix_ptr)
 		return ;
